Fix LXVST compile.
git-svn-id: svn://localhost/ardour2/branches/3.0@10188 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/vstfxinfofile.cc
@@ -283,9 +283,8 @@ static VSTFXInfo* vstfx_info_from_plugin(VSTFX *vstfx)
 /* A simple 'dummy' audiomaster callback which should be ok,
 we will only be instantiating the plugin in order to get its info*/
 
-static long simple_master_callback(struct AEffect *, int32_t opcode, int32_t, intptr_t, void *, float)
+static intptr_t simple_master_callback(struct AEffect *, int32_t opcode, int32_t, intptr_t, void *, float)
 {
-
 	if (opcode == audioMasterVersion)
 		return 2;
 	else
